added RL struct for batch update

+++ src/afl-fuzz-rl.c
@@ -105,8 +105,8 @@ void update_queue(rl_params_t *rl_params) {
 
   memcpy(send_data.data_buff, msg_array, BUFF_SIZE_SENDER * sizeof(u32));
   if (-1 == msgsnd(msqid_sender, &send_data, sizeof(t_send_u32_data) - sizeof(long), 0)) {
-  perror("msgsnd() failed");
-  exit(1);
+    perror("msgsnd() failed");
+    exit(1);
   }
 
   u32 index = 0;
@@ -114,9 +114,9 @@ void update_queue(rl_params_t *rl_params) {
     
     for (u32 i = 0; i < (BUFF_SIZE_SENDER - 1); i++) {
       if (index+i < rl_params->map_size) {
-        msg_array[i] = (u32) rl_params->positive_reward[index+i];
+        msg_array[i] = rl_params->positive_reward[index+i];
       } else {
-         msg_array[i] = (u32) 0;
+         msg_array[i] = 0;
       }
     }
     memcpy(send_data.data_buff, msg_array, BUFF_SIZE_SENDER * sizeof(u32));
@@ -132,9 +132,9 @@ void update_queue(rl_params_t *rl_params) {
     
     for (u32 i = 0; i < (BUFF_SIZE_SENDER - 1); i++) {
       if (index+i < rl_params->map_size) {
-        msg_array[i] = (u32) rl_params->negative_reward[index+i];
+        msg_array[i] = rl_params->negative_reward[index+i];
       } else {
-         msg_array[i] = (u32) 0;
+         msg_array[i] = 0;
       }
     }
     memcpy(send_data.data_buff, msg_array, BUFF_SIZE_SENDER * sizeof(u32));
